pylint refactoring
